Clamp <length> in InterpolableLength::CreateLength and add DCHECK in Length constructors, add clamping in cubic-bezier
According to spec [1], the infinity and NaN should be clamped before other components consume the values.

Therefore, the infinity and NaN values clamped from InterpolableLength to  Length which will be consumed to interpolate length.

And add DCHECK(infinite) statement in Length constructors to ensure and check the no infinity and NaN values.

The calculation of CubicBezier can be degenerated values.
Therefore add a clamp function ToFinite in calculation functions CubicBezier can return inf or -inf or NaN.

The Spec: [1]
Design docs: [2]

[1]  https://drafts.csswg.org/css-values/#calc-type-checking
[2] https://bit.ly/349gXjq

Bug: 1133390
Change-Id: Ibb7dbee997d883968c855754ab93f184c3f09306